refactor: separate match from arcade
